Fix segmentation fault in audio_hw
During the media stress cts test a segmentation fault was triggered
in audio_route.c. This was traced to the realloc call in 'path_create'
and shows a heap corruption issue.

This situation is only possible if 'audio_route_free' is called before
'audio_route_init' finishes execution (this way it is possible that
the 'ar' structure is freed and audio_route_init tries to use it at
the same time). Add locks to protect these calls, so that they do not
execute concurrently.

Added NULL assignments after pointer free calls and also make sure to
correctly free the entire audio_route structure in order to avoid memory
leaks.

Issue: APDEV-1129
Change-Id: I5c1c0f0868c51f4d3160e9c5170026a138eb9fe4
Signed-off-by: Adriana Reus <adriana.reus@intel.com>
Signed-off-by: Ravi <abhishek.ravi@intel.com>