Mic muted when headset is plugged in during record
when a headset is plugged in, the input device changes from
AUDIO_DEVICE_IN_BUILTIN_MIC to AUDIO_DEVICE_IN_WIRED_HEADSET.
When this change set through setparameters, the in_set_parameters
of audio HAL is called. However, in select_devices, value of
headset_on is set based on out_device and not in_device. Hence
in the above situation both main_mic_on and headset_on are
not set and thus the "Capture Switch" mixer setting is not
set to "On". Hence making headset_on when either input or
output device is WIRED_HEADSET.

Issue: APDEV-1129
Change-Id: I01724a11864d57ea01b57cc139c589911c44d4c0
Signed-off-by: Ravi <abhishek.ravi@intel.com>